+++ lib-fileTransfer/src/main/resources/top/xinsin/dao/FileDownloadMapper.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="top.xinsin.dao.FileUpLoadMapper">
-    <insert id="addFile" parameterType="top.xinsin.pojo.FileObject">
-        insert into file set size=#{size},name=#{name},md5=#{md5};
-    </insert>
+<mapper namespace="top.xinsin.dao.FileDownloadMapper">
     <select id="getFileName" parameterType="top.xinsin.pojo.FileObject" resultType="java.lang.String">
         select name from file where md5=#{md5};
     </select>
