+++ lib-fileTransfer/src/main/resources/top/xinsin/dao/FileUploadMapper.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
+<mapper namespace="top.xinsin.dao.FileUploadMapper">
+    <insert id="addFile" parameterType="top.xinsin.pojo.FileObject">
+        insert into file set size=#{size},name=#{name},md5=#{md5};
+    </insert>
+</mapper>
